modules/secret_manager/lambda_rotation/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.10
requests==2.33.0
